% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{as.OptPos}
\alias{as.OptPos}
\title{Coerce an argument to \code{OptPos} class.}
\usage{
as.OptPos(o = Opt(), Pos = c("Long", "Short"), Prem = 0)
}
\arguments{
\item{o}{A \code{Opt} or \code{OptPx} object}

\item{Pos}{Specify position direction in your portfolio. \code{Long} indicates that you own security (it's an asset). \code{Short} that you shorted (short sold) security (it's a liability).}

\item{Prem}{Option premium, i.e. cost of an option purchased or to be purchased.}
}
\value{
An object of class \code{OptPos}.
}
\description{
Coerce an argument to \code{OptPos} class.
}
\examples{
as.OptPos(Opt())
}
\author{
Oleg Melnikov
}

