% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{is.Opt}
\alias{is.Opt}
\title{Is an object \code{Opt}?}
\usage{
is.Opt(o)
}
\arguments{
\item{o}{Any object}
}
\value{
TRUE if and only if an argument is of \code{Opt} class.
}
\description{
Tests the argument for the specific class type.
}
\examples{
is.Opt(Opt())  #verifies that Opt() returns an object of class \\code{Opt}
is.Opt(1:3)    #verifies that code{1:3} is not an object of class \\code{Opt}
}
\author{
Oleg Melnikov
}

