% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{is.OptPos}
\alias{is.OptPos}
\title{Is an object \code{OptPos}?}
\usage{
is.OptPos(o)
}
\arguments{
\item{o}{Any object}
}
\value{
TRUE if and only if an argument is of \code{OptPos} class.
}
\description{
Tests the argument for the specific class type.
}
\examples{
is.OptPos(OptPos())
}
\author{
Oleg Melnikov
}

