% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{is.OptPx}
\alias{is.OptPx}
\title{Is an object \code{OptPx}?}
\usage{
is.OptPx(o)
}
\arguments{
\item{o}{Any object}
}
\value{
TRUE if and only if an argument is of \code{OptPx} class.
}
\description{
Tests the argument for the specific class type.
}
\examples{
is.OptPx(OptPx(Opt(S0=20), r=0.12))
}
\author{
Oleg Melnikov
}

