% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{OptPos}
\alias{OptPos}
\title{\code{OptPos} object constructor}
\usage{
OptPos(o = Opt(), Pos = c("Long", "Short"), Prem = 0)
}
\arguments{
\item{o}{An object of class \code{Opt}s}

\item{Pos}{A position direction (to the holder) with values \code{Long} for owned option contract and \code{Short} for shorted contract.}

\item{Prem}{A option premim (i.e. market cost or price), a non-negative amount to be paid for the option contract being modeled.}
}
\value{
A list of class \code{OptPx}
}
\description{
S3 object constructor for lattice-pricing specs of an option contract. Inherits \code{Opt} object.
}
\examples{
OptPos()  # Creates an S3 object for an option contract
OptPos(Opt(Right='Put'))  #See J.C.Hull, OFOD'2014, 9-ed, Fig.13.10, p.289
}
\author{
Oleg Melnikov, Department of Statistics, Rice University, Spring 2015
}

