% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QFRM.R
\name{Profit}
\alias{Profit}
\title{Computes payout/profit values}
\usage{
Profit(o = OptPos(), S = o$S0)
}
\arguments{
\item{o}{An object of class \code{Opt*}}

\item{S}{A (optional) vector or value of stock price(s) (double) at which to compute profits}
}
\value{
A numeric matrix of size \code{[length(S), 2]}. Columns: stock prices, corresponding option profits
}
\description{
Computes payout/profit values
}
\examples{
Profit(o=Opt())
graphics::plot( print( Profit(OptPos(Prem=2.5), S=40:60)), type='l'); grid()
}
\author{
Oleg Melnikov, Department of Statistics, Rice University, Spring 2015
}

